sli-csv-report-generator/requirements.txt
azure-communication-email
azure-identity
requests
pandas
openpyxl